requirements.txt
sqlalchemy==1.4.37
aiogram==2.20
PyYAML~=6.0
python-dotenv==0.20.0
psycopg2-binary==2.9.3
pandas~=1.4.3
openpyxl==3.0.10